+++ Classes/Domain/Repository/LogRepository.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Slavlee\Waf\Domain\Repository;
+
+use TYPO3\CMS\Extbase\Persistence\Repository;
+
+/*
+ * This file is part of the "waf" Extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+class LogRepository extends Repository {}